refactor(scripts): refactor dev script to copy contracts and use deterministic deployment
