+++ scripts/dev.sh
@@ -12,7 +12,6 @@ cleanup() {
     kill -9 $hardhat_pid
   fi
 }
-mnemonic="cream core pear sure dinner indoor citizen divorce sudden captain subject remember"
 
 hardhat_running() {
   nc -z localhost 8545
@@ -20,9 +19,13 @@ hardhat_running() {
 
 start-hardhat_node() {
 
+  rm -r contracts
+  cp -r node_modules/dxdao-contracts/contracts ./contracts
+  rm -r contracts/utils/GnosisSafe
+
   yarn hardhat compile
   
-  npx hardhat node > /dev/null &
+  yarn hardhat node --no-deploy > /dev/null &
 
   hardhat_pid=$!
 
@@ -43,34 +46,6 @@ fi
 echo "Starting our own hardhat node instance"
 start-hardhat_node
 
-# Compile your contracts
-yarn hardhat compile
-
-# Disable isolatedModules and use commonjs in tsconfig
-contents="$(jq '.compilerOptions.isolatedModules = false' tsconfig.json)" && \
-echo "${contents}" > tsconfig.json
-contents="$(jq '.compilerOptions.module = "commonjs"' tsconfig.json)" && \
-echo "${contents}" > tsconfig.json
-
-node scripts/beforeBuild.js
-
-# Deploy local contracts
-yarn hardhat run --network localhost scripts/dev.ts
-
-# Enable isolatedModules and use esnext as module in tsconfig
-contents="$(jq '.compilerOptions.isolatedModules = true' tsconfig.json)" && \
-echo "${contents}" > tsconfig.json
-contents="$(jq '.compilerOptions.module = "esnext"' tsconfig.json)" && \
-echo "${contents}" > tsconfig.json
-
-# Run dapp with localhost contracts
-export REACT_APP_GIT_SHA=$(echo $(git rev-parse  HEAD) | cut -c1-9)
-export SKIP_PREFLIGHT_CHECK=true
-export REACT_APP_VERSION=$npm_package_version
-
-if [[ $* == *--no-browser* ]]; then
-    echo "Setting BROWSER=none. No browser window will pop up"
-    export BROWSER=none
-fi
+DEPLOY_SALT="0x597fe520ef4c65015e0532273879fe955f79a4ff059b4f30d40aff43dda27a45" yarn hardhat --network localhost run scripts/deployGuilds.js
 
 FORCE_COLOR=true GENERATE_SOURCEMAP=false yarn react-app-rewired start | cat
